+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+### Changed
+- Updated README
+- Fixed missing AccessoryInformation.Name warning
+
+### Removed
+
 
 ## [1.7.1] - 2020-04-23