+++ CHANGELOG.md
@@ -8,12 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Extended the channelButtonService with an optional name
+- Store launch points (apps, inputs)
+- Store channel list
 
 ### Changed
 - Updated README
 - Fixed missing AccessoryInformation.Name warning
 - Improved logging
 - The channelButtonService now also accepts an array of objects as value
+- Wait till all tv information is retrieved
 
 ### Removed
 - Input names are no longer beeing saved in a file, renaming should be done in the config.json